+++ app/components/science/index.ts
@@ -1,5 +1,5 @@
 export { default as ScienceHero } from "./ScienceHero";
-export { default as ScienceQuote } from "./ScienceQuote";
+export { default as TwoSystemModel } from "./TwoSystemModel";
 export { default as ScienceAdaptogens } from "./ScienceAdaptogens";
 export { default as SciencePillars } from "./SciencePillars";
 export { default as PillarCard } from "./PillarCard";
@@ -7,4 +7,3 @@ export { default as SynergyChart } from "./SynergyChart";
 export { default as FlowVsClear } from "./FlowVsClear";
 export { default as ScienceDifferent } from "./ScienceDifferent";
 export { default as EvidenceSummary } from "./EvidenceSummary";
-export { default as SciencePageContent } from "./SciencePageContent";